add DataView, redesign BufBuilder
DataOutput is not used now. It's used incorrectly in the last commit.
DataOutput is used for formatting data to string rather than to
bytes. DataView offers interfaces to store numbers in little endian.

DataView is originally designed like DataOutput to be an abstract
class, and the subclasses inherited from it must implement
function appendBuffer. However, DataView actually needs the
information of buffer, so DataView is designed to used with a
given buffer, without the consideration about indexes being
out of bound.